+++ packages/storybook-addon-chaos/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-chaos@1.0.1...@pluralsight/ps-design-system-storybook-addon-chaos@1.0.2) (2020-03-11)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-storybook-addon-chaos
+
+
+
+
+
+# [1.0.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-chaos@0.0.25...@pluralsight/ps-design-system-storybook-addon-chaos@1.0.0) (2020-03-11)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-storybook-addon-chaos
+
+
+
+
+
 ## [0.0.25](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-chaos@0.0.24...@pluralsight/ps-design-system-storybook-addon-chaos@0.0.25) (2020-02-26)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-storybook-addon-chaos