+++ packages/storybook-addon-featureflags/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-featureflags@2.0.1...@pluralsight/ps-design-system-storybook-addon-featureflags@2.0.2) (2020-03-11)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-storybook-addon-featureflags
+
+
+
+
+
+# [2.0.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-featureflags@1.1.14...@pluralsight/ps-design-system-storybook-addon-featureflags@2.0.0) (2020-03-11)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-storybook-addon-featureflags
+
+
+
+
+
 ## [1.1.14](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-featureflags@1.1.13...@pluralsight/ps-design-system-storybook-addon-featureflags@1.1.14) (2020-02-18)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-storybook-addon-featureflags